Fix systemd targets
For one, this should not be installed into the graphical target, as its main
purpose is indeed to set up headless systems where the graphical.target
may not even be started.

Secondly, having set After=multi-user.target may cause it to not start
at all if another broken service in the multi-user.target fails to start
and thus the multi-user.target may not complete.